+++ src/styles/todoapp.scss
@@ -56,7 +56,7 @@
   }
 
   &__new-todo {
-    width: 100%;
+    width: 85%;
     padding: 16px 16px 16px 60px;
 
     font-size: 24px;
